Remove .vscode directory generation
