+++ compiler-plugin-tests/src/test/java/org/ballerinax/azurefunctions/test/DeploymentTest.java
@@ -84,7 +84,6 @@ public class DeploymentTest {
         Assert.assertEquals(defaultWorkerPath, "main.jar");
         Files.deleteIfExists(depedenciesToml);
         Files.deleteIfExists(handlers.resolve("main.jar"));
-        FileUtils.deleteDirectory(handlers.resolve(".vscode").toFile());
         FileUtils.deleteDirectory(handlers.resolve("azure_functions").toFile());
     }
 }
