--- compiler-plugin/src/main/java/org/ballerinax/azurefunctions/tooling/LocalSettings.java
@@ -1,44 +1,0 @@
-/*
- * Copyright (c) 2022, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.ballerinax.azurefunctions.tooling;
-
-import com.google.gson.annotations.SerializedName;
-
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- * Represents local.settings.json in the target directory.
- *
- * @since 2201.3.0
- */
-public class LocalSettings {
-
-    @SerializedName("IsEncrypted")
-    private boolean isEncrypted;
-
-    @SerializedName("Values")
-    private Map<String, String> values;
-
-    public LocalSettings() {
-        this.isEncrypted = false;
-        this.values = new HashMap<>();
-        this.values.put("AzureWebJobsStorage", "");
-        this.values.put("FUNCTIONS_WORKER_RUNTIME", "java");
-    }
-}
