--- compiler-plugin/src/main/java/org/ballerinax/azurefunctions/tooling/Settings.java
@@ -1,53 +1,0 @@
-/*
- * Copyright (c) 2022, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.ballerinax.azurefunctions.tooling;
-
-import com.google.gson.annotations.SerializedName;
-import io.ballerina.projects.Project;
-import org.ballerinax.azurefunctions.Util;
-
-/**
- * Represents settings.json in the .vscode directory.
- *
- * @since 2201.3.0
- */
-public class Settings {
-
-    @SerializedName("azureFunctions.deploySubpath")
-    private String deploySubpath;
-
-    @SerializedName("azureFunctions.projectLanguage")
-    private String projectLanguage;
-
-    @SerializedName("azureFunctions.projectRuntime")
-    private String projectRuntime;
-
-    @SerializedName("debug.internalConsoleOptions")
-    private String internalConsoleOptions;
-
-    @SerializedName("azureFunctions.projectSubpath")
-    private String projectSubpath;
-
-    public Settings(Project project) {
-        this.deploySubpath = Util.getAzureFunctionsRelative(project);
-        this.projectLanguage = "Custom";
-        this.projectRuntime = "~4";
-        this.internalConsoleOptions = "neverOpen";
-        this.projectSubpath = Util.getAzureFunctionsRelative(project);
-    }
-}
